+++ tests/index.node.js
@@ -17,6 +17,7 @@ import * as customCompare from './custom-compare.test.js'
 import * as inlineNodes from './inline-nodes.test.js'
 import * as attrAttribution from './attr-attribution.test.js'
 import * as ySyncRdt from './y-sync-rdt.test.js'
+import * as prosemirrorRdt from './prosemirror-rdt.test.js'
 // import * as tr from './tr.test.js'
 
 import { runTests } from 'lib0/testing'
@@ -71,7 +72,8 @@ runTests({
   customCompare,
   inlineNodes,
   attrAttribution,
-  ySyncRdt
+  ySyncRdt,
+  prosemirrorRdt
   // prosemirror,
   // tr
 }).then(success => {